requirements.txt
dgl==0.9.1
diskcache==5.6.3
einops==0.8.0
numpy==2.0.1
pandas==2.2.2
pytz==2024.1
PyYAML==6.0.1
PyYAML==6.0.1
scikit_learn==1.4.2
torch==2.3.0
torch_geometric==2.5.3
tqdm==4.66.2
